fix: create separate category page header component
