+++ packages/core/nuxt-theme-module/theme/pages/Category.vue
@@ -13,77 +13,7 @@
             class="navbar__title" />
         </LazyHydrate>
       </div>
-
-      <div class="navbar__main">
-
-        <!-- Filter and sort buttons are disabled until building seperate component and apply filter and sort logic -->
-
-        <!-- <LazyHydrate on-interaction>
-          <SfButton
-            class="sf-button--text navbar__filters-button"
-            aria-label="Filters"
-            @click="toggleFilterSidebar"
-          >
-            <SfIcon
-              size="24px"
-              color="dark-secondary"
-              icon="filter2"
-              class="navbar__filters-icon"
-            />
-            {{ $t('Filters') }}
-          </SfButton>
-        </LazyHydrate>
-
-        <div class="navbar__sort desktop-only">
-          <span class="navbar__label">{{ $t('Sort by') }}:</span>
-          <LazyHydrate on-interaction>
-            <SfSelect
-              :value="sortBy.selected"
-              placeholder="Select sorting"
-              class="navbar__select"
-              @input="th.changeSorting"
-            >
-              <SfSelectOption
-                v-for="option in sortBy.options"
-                :key="option.id"
-                :value="option.id"
-                class="sort-by__option"
-                >{{ option.value }}</SfSelectOption
-              >
-            </SfSelect>
-          </LazyHydrate>
-        </div> -->
-
-        <div class="navbar__counter">
-          <span class="navbar__label desktop-only">{{ $t('Products found') }}: </span>
-          <span class="desktop-only">{{ pagination.totalItems }}</span>
-          <span class="navbar__label smartphone-only">{{ pagination.totalItems }} {{ $t('Items') }}</span>
-        </div>
-
-        <div class="navbar__view">
-          <span class="navbar__view-label desktop-only">{{ $t('View') }}</span>
-          <SfIcon
-            class="navbar__view-icon"
-            :color="isCategoryGridView ? 'black' : 'dark-secondary'"
-            icon="tiles"
-            size="12px"
-            role="button"
-            aria-label="Change to grid view"
-            :aria-pressed="isCategoryGridView"
-            @click="toggleCategoryGridView"
-          />
-          <SfIcon
-            class="navbar__view-icon"
-            :color="!isCategoryGridView ? 'black' : 'dark-secondary'"
-            icon="list"
-            size="12px"
-            role="button"
-            aria-label="Change to list view"
-            :aria-pressed="!isCategoryGridView"
-            @click="toggleCategoryGridView"
-          />
-        </div>
-      </div>
+      <CategoryPageHeader :pagination="pagination"/>
     </div>
 
     <div class="main section">
@@ -253,12 +183,6 @@
         </div>
       </SfLoader>
     </div>
-
-    <!-- Filters sidebar is disabled until filter and sort logic is applied -->
-
-    <!-- <LazyHydrate when-idle>
-      <FiltersSidebar @close="toggleFilterSidebar"/>
-    </LazyHydrate>         -->
   </div>
 </template>
 
@@ -276,20 +200,38 @@ import {
   SfAccordion,
   SfSelect,
   SfBreadcrumbs,
-  SfLoader,  
+  SfLoader,
   SfProperty
 } from '@storefront-ui/vue';
+import CategoryPageHeader from '../components/CategoryPageHeader';
 import { computed } from '@vue/composition-api';
 import { useCart, useWishlist, productGetters, useFacet, facetGetters } from '<%= options.generate.replace.composables %>';
 import { useUiHelpers, useUiState } from '~/composables';
 import { onSSR } from '@vue-storefront/core';
 import LazyHydrate from 'vue-lazy-hydration';
-import FiltersSidebar from "../components/FiltersSidebar"
 
 // TODO(addToCart qty, horizontal): https://github.com/vuestorefront/storefront-ui/issues/1606
 export default {
   transition: 'fade',
-  setup(props, context) {
+  components: {
+    SfButton,
+    SfIcon,
+    SfList,
+    SfFilter,
+    SfProductCard,
+    SfProductCardHorizontal,
+    SfPagination,
+    SfMenuItem,
+    SfAccordion,
+    SfSelect,
+    SfBreadcrumbs,
+    SfLoader,
+    SfHeading,
+    SfProperty,
+    LazyHydrate,
+    CategoryPageHeader
+  },
+  setup() {
     const th = useUiHelpers();
     const uiState = useUiState();
     const { addItem: addItemToCart, isInCart } = useCart();
@@ -299,8 +241,6 @@ export default {
     const products = computed(() => facetGetters.getProducts(result.value));
     const categoryTree = computed(() => facetGetters.getCategoryTree(result.value));
     const breadcrumbs = computed(() => facetGetters.getBreadcrumbs(result.value));
-    const sortBy = computed(() => facetGetters.getSortOptions(result.value));
-    const facets = computed(() => facetGetters.getGrouped(result.value, ['color', 'size']));
     const pagination = computed(() => facetGetters.getPagination(result.value));
     const activeCategory = computed(() => {
       const items = categoryTree.value.items;
@@ -327,31 +267,11 @@ export default {
       productGetters,
       pagination,
       activeCategory,
-      sortBy,
-      facets,
       breadcrumbs,
       addItemToWishlist,
       addItemToCart,
-      isInCart,
+      isInCart
     };
-  },
-  components: {
-    SfButton,
-    SfIcon,
-    SfList,
-    SfFilter,
-    SfProductCard,
-    SfProductCardHorizontal,
-    SfPagination,
-    SfMenuItem,
-    SfAccordion,
-    SfSelect,
-    SfBreadcrumbs,
-    SfLoader,    
-    SfHeading,
-    SfProperty,
-    LazyHydrate,
-    FiltersSidebar
   }
 };
 </script>
@@ -376,147 +296,38 @@ export default {
   margin: var(--spacer-base) auto var(--spacer-lg);
 }
 .navbar {
-  position: relative;
-  display: flex;
-  border: 1px solid var(--c-light);
-  border-width: 0 0 1px 0;
-  @include for-desktop {
-    border-width: 1px 0 1px 0;
-  }
-  &.section {
-    padding: var(--spacer-sm);
-    @include for-desktop {
-      padding: 0;
-    }
-  }
-  &__aside,
-  &__main {
+    position: relative;
     display: flex;
-    align-items: center;
-    padding: var(--spacer-sm) 0;
-  }
-  &__aside {
-    flex: 0 0 15%;
-    padding: var(--spacer-sm) var(--spacer-sm);
     border: 1px solid var(--c-light);
-    border-width: 0 1px 0 0;
-  }
-  &__main {
-    flex: 1;
-    padding: 0;
-    justify-content: space-between;
+    border-width: 0 0 1px 0;
     @include for-desktop {
-      padding: var(--spacer-xs) var(--spacer-xl);
+      border-width: 1px 0 1px 0;
     }
-  }
-  &__title {
-    --heading-title-font-weight: var(--font-weight--semibold);
-    --heading-title-font-size: var(--font-size--xl);
-  }
-  &__filters-icon {
-    margin: 0 0 0 var(--spacer-xs);
-    order: 1;
-    @include for-desktop {
-      margin: 0 var(--spacer-xs) 0 0;
-      order: 0;
-    }
-  }
-  &__filters-button {
-    display: flex;
-    align-items: center;
-    --button-font-size: var(--font-size--base);
-    --button-text-decoration: none;
-    --button-color: var(--c-link);
-    --button-font-weight: var(--font-weight--normal);
-    @include for-mobile {
-      --button-font-weight: var(--font-weight--medium);
-      order: 2;
-    }
-    svg {
-      fill: var(--c-text-muted);
-      transition: fill 150ms ease;
-    }
-    &:hover {
-      svg {
-        fill: var(--c-primary);
+    &.section {
+      padding: var(--spacer-sm);
+      @include for-desktop {
+        padding: 0;
       }
     }
-  }
-  &__label {
-    font-family: var(--font-family--secondary);
-    font-weight: var(--font-weight--normal);
-    color: var(--c-text-muted);
-    @include for-desktop {
-      color: var(--c-link);
-      margin: 0 var(--spacer-2xs) 0 0;
+    &__aside,
+    {
+      display: flex;
+      align-items: center;
+      padding: var(--spacer-sm) 0;
+    }
+    &__aside {
+      flex: 0 0 15%;
+      padding: var(--spacer-sm) var(--spacer-sm);
+      border: 1px solid var(--c-light);
+      border-width: 0 1px 0 0;
+    }
+    &__title {
+      --heading-title-font-weight: var(--font-weight--semibold);
+      --heading-title-font-size: var(--font-size--xl);
     }
   }
-  &__select {
-    --select-width: 220px;
-    --select-padding: 0;
-    --select-height: auto;
-    --select-selected-padding: 0 var(--spacer-lg) 0 var(--spacer-2xs);
-    --select-margin: 0;
-    --select-option-font-size: var(--font-size-sm);
-    --select-error-message-height: 0;
-    ::v-deep .sf-select__dropdown {
-      font-size: var(--font-size-sm);
-      font-family: var(--font-family--secondary);
-      font-weight: var(--font-weight--light);
-      margin: 0;
-    }
-    ::v-deep .sf-select__placeholder {
-      --select-option-font-size: var(--font-size-sm);
-    }
-  }
-  &__sort {
-    display: flex;
-    align-items: center;
-    margin: 0 auto 0 var(--spacer-2xl);
-  }
-  &__counter {
-    font-family: var(--font-family--secondary);
-    order: 1;
-    @include for-desktop {
-      order: 0;
-    }
-  }
-  &__view {
-    display: flex;
-    align-items: center;
-    order: 0;
-    @include for-desktop {
-      margin: 0 0 0 var(--spacer-2xl);
-      order: 0;
-    }
-    &-icon {
-      cursor: pointer;
-      margin: 0 var(--spacer-base) 0 0;
-      &:last-child {
-        margin: 0;
-      }
-    }
-    &-label {
-      margin: 0 var(--spacer-sm) 0 0;
-      font: var(--font-weight--normal) var(--font-size--base) / 1.6
-        var(--font-family--secondary);
-      text-decoration: none;
-      color: var(--c-link);
-    }
-  }
-}
-.sort-by {
-  flex: unset;
-  width: 11.875rem;
-}
 .main {
   display: flex;
-}
-.sidebar {
-  flex: 0 0 15%;
-  padding: var(--spacer-sm);
-  border: 1px solid var(--c-light);
-  border-width: 0 1px 0 0;
 }
 .list {
   --menu-item-font-size: var(--font-size--sm);
@@ -601,6 +412,12 @@ export default {
       font-size: var(--font-size--sm);
     }
   }
+}
+.sidebar {
+  flex: 0 0 15%;
+  padding: var(--spacer-sm);
+  border: 1px solid var(--c-light);
+  border-width: 0 1px 0 0;
 }
 .loading {
   margin: var(--spacer-3xl) auto;
